+++ lib/src/directory/directory.dart
@@ -1,0 +1,3 @@
+export 'unsupported.dart'
+    if (dart.library.html) 'web.dart'
+    if (dart.library.io) 'io.dart';
